+++ src/main/java/org/moon/figura/gui/widgets/lists/ConfigList.java
@@ -115,7 +115,7 @@ public class ConfigList extends AbstractList {
 
     public void updateScroll() {
         //store old scroll pos
-        float pastScroll = (totalHeight - height) * scrollBar.getScrollProgress();
+        double pastScroll = (totalHeight - height) * scrollBar.getScrollProgress();
 
         //get new height
         totalHeight = -4;
