+++ src/main/java/org/moon/figura/gui/widgets/lists/PlayerList.java
@@ -175,7 +175,7 @@ public class PlayerList extends AbstractList {
 
     public void updateScroll() {
         //store old scroll pos
-        float pastScroll = (totalHeight - height) * scrollBar.getScrollProgress();
+        double pastScroll = (totalHeight - height) * scrollBar.getScrollProgress();
 
         //get new height
         totalHeight = 0;
